Add files via upload
The Celsius to Fahrenheit Converter is a simple yet handy web application that allows users to convert temperature values from Celsius to Fahrenheit and vice versa. This project is a practical example of a web application built with HTML, CSS, and JavaScript.
Technology Stack:

HTML: Used to structure the web page, including input fields and buttons.

CSS: Provides styling to make the application visually appealing and responsive.

JavaScript: Powers the conversion logic and enables real-time updates without the need for page reloads.